requirements.txt
streamlit
fastapi
uvicorn
tensorflow
scikit-learn
numpy
joblib
pillow
requests
